+++ src/main/java/services/sandbox/model/PriceAlert.java
@@ -1,6 +1,7 @@
 package services.sandbox.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.Instant;
 
 /**
@@ -11,7 +12,7 @@ public class PriceAlert implements Serializable {
     private String id;
     private String userId;
     private String ticker;
-    private double targetPrice;
+    private BigDecimal targetPrice;
     /** true = notify when price >= target, false = notify when price <= target */
     private boolean above;
     private Instant createdAt;
@@ -20,7 +21,7 @@ public class PriceAlert implements Serializable {
     }
 
     public PriceAlert(String id, String userId, String ticker,
-                      double targetPrice, boolean above, Instant createdAt) {
+                      BigDecimal targetPrice, boolean above, Instant createdAt) {
         this.id = id;
         this.userId = userId;
         this.ticker = ticker;
@@ -32,7 +33,7 @@ public class PriceAlert implements Serializable {
     public String getId() { return id; }
     public String getUserId() { return userId; }
     public String getTicker() { return ticker; }
-    public double getTargetPrice() { return targetPrice; }
+    public BigDecimal getTargetPrice() { return targetPrice; }
     public boolean isAbove() { return above; }
     public Instant getCreatedAt() { return createdAt; }
 }
